requirements.txt
numpy==1.24.3
Flask==2.2.2
Werkzeug==2.3.7
deepface==0.0.79
faiss-cpu==1.7.4
prisma==0.11.0
sanic==23.6.0
ujson==5.8.0
Pyrebase4==4.7.1


